restructured json serialization, merged all instances of json options
