+++ src/CSharp App/Json/Converters/NormalizeZeroConverter.cs
@@ -1,7 +1,7 @@
-using Newtonsoft.Json;
 using System;
+using Newtonsoft.Json;
 
-namespace VolumetricSelection2077.Json;
+namespace VolumetricSelection2077.Json.Converters;
 
 public class NormalizeZeroConverter : JsonConverter
 {
