+++ src/CSharp App/TestingStuff/TestWheezeKitSerialization.cs
@@ -1,10 +1,9 @@
 using System.IO;
 using Newtonsoft.Json;
-using VolumetricSelection2077.Json;
+using VolumetricSelection2077.Json.Helpers;
 using VolumetricSelection2077.Models;
 using VolumetricSelection2077.Models.WorldBuilder.Editor;
-using VolumetricSelection2077.Models.WorldBuilder.Spawn;
-using VolumetricSelection2077.Models.WorldBuilder.Spawn.Mesh;
+
 using VolumetricSelection2077.Services;
 
 namespace VolumetricSelection2077.TestingStuff;
@@ -30,17 +29,11 @@ public class TestWheezeKitSerialization : IDebugTool
         var testFilePath = @"E:\Games\Cyberpunk 2077\bin\x64\plugins\cyber_engine_tweaks\mods\entSpawner\data\objects\Huge Penthouse Mockup.json";
         var testFileContent = File.ReadAllText(testFilePath);
         
-        var options = new JsonSerializerSettings 
-        {
-            Converters = { new WorldBuilderElementJsonConverter(), new WorldBuilderSpawnableJsonConverter()},
-            Formatting = Formatting.Indented
-        };
-        
-        var deserializedElement = JsonConvert.DeserializeObject<Element>(testFileContent, options);
+        var deserializedElement = JsonConvert.DeserializeObject<Element>(testFileContent, JsonSerializerPresets.WorldBuilder);
         
         
         LogResourcePathsRecursively(deserializedElement);
         
-        Logger.Debug($"Deserialized element:\n{JsonConvert.SerializeObject(deserializedElement, options)}");
+        Logger.Debug($"Deserialized element:\n{JsonConvert.SerializeObject(deserializedElement, JsonSerializerPresets.WorldBuilder)}");
     }
 }
